🦄 refactor(msgbus): move alert struct to prometheus pkg
