+++ pkg/msgbus/switcher/alert.go
@@ -24,38 +24,10 @@ import (
 	"kubegems.io/kubegems/pkg/utils"
 	"kubegems.io/kubegems/pkg/utils/database"
 	"kubegems.io/kubegems/pkg/utils/prometheus"
+	"kubegems.io/kubegems/pkg/utils/set"
 )
 
-type WebhookAlert struct {
-	Receiver          string            `json:"receiver"`
-	Status            string            `json:"status"`
-	Alerts            []Alert           `json:"alerts"`
-	GroupLabels       map[string]string `json:"groupLabels"`
-	CommonLabels      map[string]string `json:"commonLabels"`
-	CommonAnnotations map[string]string `json:"commonAnnotations"`
-	ExternalURL       string            `json:"externalURL"`
-	Version           string            `json:"version"`
-	GroupKey          string            `json:"groupKey"`
-	TruncatedAlerts   int64             `json:"truncatedAlerts"`
-	DataBase          *database.Database
-}
-
-func (w *WebhookAlert) fingerprintMap() map[string][]Alert {
-	ret := map[string][]Alert{}
-	for _, v := range w.Alerts {
-		alerts, ok := ret[v.Fingerprint]
-		if ok {
-			alerts = append(alerts, v)
-		} else {
-			alerts = []Alert{v}
-		}
-		ret[v.Fingerprint] = alerts
-	}
-
-	return ret
-}
-
-func (ms *MessageSwitcher) saveFingerprintMapToDB(fingerprintMap map[string][]Alert) []models.AlertMessage {
+func (ms *MessageSwitcher) saveFingerprintMapToDB(fingerprintMap map[string][]prometheus.Alert) []models.AlertMessage {
 	now := time.Now()
 	alertMessages := []models.AlertMessage{}
 	alertInfos := []models.AlertInfo{}
@@ -106,24 +78,6 @@ func (ms *MessageSwitcher) saveFingerprintMapToDB(fingerprintMap map[string][]Al
 	return alertMessages
 }
 
-type Alert struct {
-	Status       string            `json:"status"`
-	Labels       map[string]string `json:"labels"`
-	Annotations  map[string]string `json:"annotations"`
-	StartsAt     *time.Time        `json:"startsAt"`
-	EndsAt       *time.Time        `json:"endsAt"`
-	GeneratorURL string            `json:"generatorURL"`
-	Fingerprint  string            `json:"fingerprint"`
-}
-
-func (a *Alert) AlertName() string {
-	return a.Labels["alertname"]
-}
-
-func (a *Alert) Detail() string {
-	return fmt.Sprintf("%s: %s", a.AlertName(), a.Annotations["message"])
-}
-
 type ResID struct {
 	ClusterID       uint
 	EnvironmentID   uint
@@ -135,21 +89,20 @@ type ResID struct {
 }
 
 // map 效率更高
-func (w *WebhookAlert) GetAlertUsers(pos database.AlertPosition) map[uint]struct{} {
-	tmp := []uint{}
+func GetAlertUsers(w *prometheus.WebhookAlert, pos database.AlertPosition, database *database.Database) *set.Set[uint] {
+	ret := set.NewSet[uint]()
 	switch w.CommonLabels["gems_alert_scope"] {
 	case prometheus.ScopeSystemAdmin:
-		tmp = append(tmp, w.DataBase.SystemAdmins()...) // 系统管理员
+		ret.Append(database.SystemAdmins()...) // 系统管理员
 	case prometheus.ScopeSystemUser:
-		tmp = append(tmp, w.DataBase.SystemUsers()...) // 系统所有用户
+		ret.Append(database.SystemUsers()...) // 系统所有用户
 	default: // normal and null
-		tmp = append(tmp, w.DataBase.EnvUsers(pos.EnvironmentID)...)  // 环境用户
-		tmp = append(tmp, w.DataBase.ProjectAdmins(pos.ProjectID)...) // 项目管理员
-		tmp = append(tmp, w.DataBase.TenantAdmins(pos.TenantID)...)   // 租户管理员
-	}
-	ret := make(map[uint]struct{})
-	for _, v := range tmp {
-		ret[v] = struct{}{}
+		// 环境用户
+		// 项目管理员
+		// 租户管理员
+		ret.Append(database.EnvUsers(pos.EnvironmentID)...).
+			Append(database.ProjectAdmins(pos.ProjectID)...).
+			Append(database.TenantAdmins(pos.TenantID)...)
 	}
 	return ret
 }
